VALIDATION complete!
From last commit tested compile and implement with jar file. Works on
Linux and Windows. Validation (VDate.java) now complete. Many classes and
methods added. Used old validation logic as a framework and found even
though it was basically sound .. there were a number of bugs and/or
wrong use of classes/implementations. Triggers used on back end to
manage the status of the file as user finishes manually checking of
anomalies. Also used Transaction Update in DbHelper.resetData()
Next to implement is writing the validated data to the database.